move choi functions in mitiq/pec/utils.py
